requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
Django==4.0.2
django-braces==1.15.0
django-embed-video==1.4.2.dev1+gf2b134c
djangorestframework==3.13.1
idna==3.3
python-memcached==1.59
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
